Change the database to use decimals, fixed some bugs.

+++ db/migrate/20130716074450_add_guidance_info_to_exercise.rb
@@ -1,9 +1,9 @@
 class AddGuidanceInfoToExercise < ActiveRecord::Migration
   def change
-    add_column :exercises, :duration_after_repetition, :integer
-    add_column :exercises, :duration_up_repetition, :integer
-    add_column :exercises, :duration_middle_repetition, :integer
-    add_column :exercises, :duration_down_repetition, :integer
+    add_column :exercises, :duration_after_repetition, :decimal
+    add_column :exercises, :duration_up_repetition, :decimal
+    add_column :exercises, :duration_middle_repetition, :decimal
+    add_column :exercises, :duration_down_repetition, :decimal
     add_column :exercises, :guidance_type, :string, :null => false, :default => 'manual'
     
     remove_column :series, :duration_after_repetition, :integer
